Implemented the methods to load and save templates through the TemplateController for use by the Template Designer.
Implemented methods to get / set the template key of a CoalesceEntityTemplate. This will default to a UUID based off a MD5 hash of the name, source, and version. This means registering the same template from an entity will maintain the same key.

Implemented a Template unit test in AbstractCoalescePersistorTest.

Implemented Template support in the Mock and File persister implementations.