Switched back to using the bundle-geomesa, modified the bundles and poms to so the IT pass within Karaf.
Reverted changes to the IT test. Its ok to change where they are pointing
in your local copy but whats committed should be localhost.

Spifly is a framework that enables service loaders within OSGi frameworks.
However I was not able to get this working due to an exception loading the
logger in the DataStoreFinder. Left the chnages in commented out so it can
be looked into again at a later time. Ideally bundle-geotools should have
no references to geomesa.