Expanded the parallel task to include the mutations.
Implemented a feature option to either use the feature or data store

Implemented ability to filter which features are to be returned by the iterator.

Updated EntityLinkHelper to use ECoalesceObjectStatus instead of the readonly flag.

Modified CoalesceComponentImpl to use the simple name if TRACE is not enabled instead of the full name.

Modified MetricResuts to store the name of the task for better reporting.