Implemented unit test to verify the format of returned columns.
Fixed the search results to return a WKT value instead of a postgres object when returning a geometry column.

Fixed issue with Accumulo using VARCHAR for all the returned columns of a search.

Fixed issue with Accumulo not handling enumerated list.

Fixed issue with Accumulo wrapping returned list with brackets.