Pulled the configuration of persisters from a property file out of the framework and into a utility class.
Modified the logic to respect the order of the persisters defined in the configuration file with the first being the authoritative.

Modified the logic so if the authoritative persister does not meet the capability of the requested action instead of throwing an error it will now use the first secondary persister capable of fulfilling the request.