Cleaned up persister capabilities and exposed it via the CoalesceSearchFramework.
Switched CoalecePropertyFactory to use FileFactor2 which has support for BBOX etc.

Implemented Derby's delete template so a unit test could pass.

Implemented Between and During unit test for the search API

Implemented ability to register templates that are already stored within a database by passing in a key. To be used within a blueprint.